refactor: split page into server wrapper + client component for dynamic title metadata
- page.tsx: server component exporting generateMetadata for dynamic page title
- SessionDetailClient.tsx: client component with all interactive logic
- Tests now import SessionDetailClient directly + test generateMetadata
- 20 tests passing
